requirements.txt
opencv-python
pytest
pillow
matplotlib
tqdm
beautifulsoup4
scikit-learn
pandas
tabulate
XlsxWriter
absl-py
imgaug
wandb
tensorflow-addons
ray[tune]
hyperopt
pytorch-lightning
lightning-bolts
tensorflow-probability
seaborn
latex
